Implement SpikeVectorSortingSegment and add zarr to core dependencies
